# api/trustgraph.openapi.yaml
openapi: 3.1.0
info:
  title: TrustGraph API
  version: 0.2.0
  description: |
    Trust, safety, and verification service for ConnectionSphere.

    Planes:
      - Plane A: first-party signals and trust assessment.
      - Plane B: consented verification (LinkedIn, government ID, liveness,
        image). Every Plane B read/write requires an active consent.
      - Plane C: investigation tools, restricted to authenticated investigators.

    Auth:
      - Plane A assessment endpoints are called service-to-service.
      - Plane B endpoints are subject-scoped and reached by the app on behalf
        of the subject; consent is enforced server-side per operation.
      - `/v1/admin/*` and `/v1/metrics/*` require `Authorization: Bearer
        $ADMIN_TOKEN`.
      - `/v1/investigations/*` require `Authorization: Bearer
        $INVESTIGATOR_TOKEN` (or `$ADMIN_TOKEN` for supervisor-level access),
        plus an optional `X-Investigator-Actor` header naming the human.

    When a vendor integration is not configured, its endpoint returns
    `503 not_configured`. It never returns a passing result it did not obtain.
servers:
  - url: http://localhost:8081
    description: Local development (via ConnectionSphere Docker network)
  - url: https://trustgraph-api.onrender.com
    description: Render production
...
paths:
...
  /v1/assessments:
    post:
      operationId: createAssessment
      summary: Request a trust assessment for a ConnectionSphere user
      tags: [Assessment]
      description: |
        Evaluates signals and returns a trust tier, risk score, and decision.
        Idempotent — resending the same idempotencyKey within 24 hours returns
        the cached result.

        The age gate runs here: a subject whose date of birth puts them under
        18 is blocked and the block is audited as `age_gate_blocked`.

        Timeout: 300ms recommended from the caller. If TrustGraph is
        unavailable, the caller should fail-open to provisional tier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentRequest"
            example:
              contractVersion: "2026-08-01"
              idempotencyKey: "reg-abc123-attempt1"
              subject:
                connectionSphereUserId: "usr_12345"
                email: "user@example.com"
                phone: "+15551234567"
                dateOfBirth: "1996-04-02"
              signals:
                emailVerified: true
                phoneVerified: false
                deviceFingerprint: "fp_a1b2c3d4"
                ipAddress: "192.168.1.100"
                imageHash: ""
              requestContext:
                userAgent: "Mozilla/5.0"
                correlationId: "corr-xyz-789"
      responses:
        "200":
          description: Assessment completed (or cached idempotent result)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Age gate blocked the subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/assessments/{assessmentId}:
    get:
      operationId: getAssessment
      summary: Retrieve a previous assessment by ID
      tags: [Assessment]
      parameters:
        - $ref: "#/components/parameters/AssessmentId"
      responses:
        "200":
          description: Assessment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "400":
          description: Invalid assessment ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/admin/reviews/{assessmentId}:
    post:
      operationId: submitReview
      summary: Record a human review decision
      tags: [Admin]
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/AssessmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              properties:
                outcome:
                  type: string
                  enum: [uphold, overturn, escalate]
                notes:
                  type: string
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    AssessmentId:
      name: assessmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or malformed bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Token invalid, or role insufficient for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ServiceAccountHeader:
      type: apiKey
      in: header
      name: X-Service-Account
      description: ConnectionSphere service account identifier
    AdminBearer:
      type: http
      scheme: bearer
      description: "Set via ADMIN_TOKEN. Grants supervisor-level Plane C access."
    InvestigatorBearer:
      type: http
      scheme: bearer
      description: "Set via INVESTIGATOR_TOKEN. Optional X-Investigator-Actor header names the human."
  schemas:
    AssessmentRequest:
      type: object
      required:
        - contractVersion
        - idempotencyKey
        - subject
      properties:
        contractVersion:
          type: string
          description: API contract version for backward compatibility
          example: "2026-08-01"
        idempotencyKey:
          type: string
          description: Unique key per request. Same key within 24h returns cached result.
          example: "reg-abc123-attempt1"
        subject:
          $ref: "#/components/schemas/SubjectData"
        signals:
          $ref: "#/components/schemas/SignalsData"
        requestContext:
          $ref: "#/components/schemas/RequestContext"
    SubjectData:
      type: object
      required:
        - connectionSphereUserId
      properties:
        connectionSphereUserId:
          type: string
          example: "usr_12345"
        email:
          type: string
          format: email
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
          description: Used by the age gate. Under-18 subjects are blocked.
    SignalsData:
      type: object
      properties:
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        deviceToken:
          type: string
        deviceFingerprint:
          type: string
        ipAddress:
          type: string
          format: ipv4
        imageHash:
          type: string
          description: Perceptual hash of profile image
    RequestContext:
      type: object
      properties:
        userAgent:
          type: string
        correlationId:
          type: string
    AssessmentResponse:
      type: object
      properties:
        contractVersion:
          type: string
        assessmentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, complete, error, deferred]
        trustTier:
          type: string
          enum: [provisional, standard, elevated, limited]
          description: |
            provisional — new user, limited capabilities until signals improve
            standard — normal user, full capabilities
            elevated — verified user, trust badges visible
            limited — restricted, pending human review
        riskBand:
          type: string
          enum: [low, elevated, high, unknown]
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Higher = riskier
        reasonCodes:
          type: array
          items:
            type: string
          description: Structured codes explaining the decision
          example:
            - EMAIL_VERIFIED
            - PHONE_NOT_VERIFIED
            - DEVICE_FIRST_SEEN
        decision:
          type: string
          enum: [accept, verify, review, deny]
        requiredActions:
          type: array
          items:
            type: string
          example:
            - VERIFY_PHONE
        policyVersion:
          type: string
          example: "registration-v1"
        completedAt:
          type: string
          format: date-time
        signals:
          type: object
          properties:
            processed:
              type: array
              items:
                type: string
            skipped:
              type: array
              items:
                type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string